+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Change `const PARTIAL_ROUNDS` to 60 (was 59) [#260]
 - Increase the number of constants to 340 (was 335) [#260]
+- Change `finalize` to return the hash directly (not wrapped in a `Result`) [#259]
 
 ## [0.37.0] - 2024-03-27
 
@@ -499,6 +500,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- ISSUES -->
 [#260]: https://github.com/dusk-network/poseidon252/issues/260
+[#259]: https://github.com/dusk-network/poseidon252/issues/259
 [#251]: https://github.com/dusk-network/poseidon252/issues/251
 [#248]: https://github.com/dusk-network/poseidon252/issues/248
 [#246]: https://github.com/dusk-network/poseidon252/issues/246